requirements.txt
# LangChain核心包
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# 数据处理
pandas==2.1.4
numpy==1.24.3

# 向量数据库
chromadb==0.4.18

# 环境变量管理
python-dotenv==1.0.0

# HTTP请求
requests==2.31.0

# 日志
loguru==0.7.2

# CORS支持
fastapi-cors==0.0.6

# 静态文件服务
aiofiles==23.2.1

# JSON处理
pydantic==2.5.0

# 文档嵌入
sentence-transformers==2.2.2

# 中文分词
jieba==0.42.1